Move GlyfPost struct and shared impls into writer module
